Update 2
I updated the home-page.php file and tried to make a new services.php
template. However, when I make a new Services page in WP, I can see the
Services template and select it, but the field don’t come up when I try
to create and publish the page.